Populate code of conduct
This CoC was also borrowed from Open Tech Strategies; the main
thing that is not ideal is the fact that the CoC team is just
one person -- this is a problem if that person violates the CoC,
and should be improved.

For now, however, this is better than absolutely nothing.